+++ src/lib/output/themes/default/DefaultThemeRenderContext.ts
@@ -12,7 +12,12 @@ import { defaultLayout } from "./layouts/default";
 import { index } from "./partials";
 import { analytics } from "./partials/analytics";
 import { breadcrumb } from "./partials/breadcrumb";
-import { comment, commentSummary, commentTags } from "./partials/comment";
+import {
+    comment,
+    commentSummary,
+    commentTags,
+    reflectionFlags,
+} from "./partials/comment";
 import { footer } from "./partials/footer";
 import { header } from "./partials/header";
 import { hierarchy } from "./partials/hierarchy";
@@ -110,6 +115,7 @@ export class DefaultThemeRenderContext {
     comment = bind(comment, this);
     commentSummary = bind(commentSummary, this);
     commentTags = bind(commentTags, this);
+    reflectionFlags = bind(reflectionFlags, this);
     footer = bind(footer, this);
     header = bind(header, this);
     hierarchy = bind(hierarchy, this);