+++ src/lib/output/themes/default/partials/member.tsx
@@ -1,4 +1,4 @@
-import { classNames, getDisplayName, renderFlags, wbr } from "../../lib";
+import { classNames, getDisplayName, wbr } from "../../lib";
 import type { DefaultThemeRenderContext } from "../DefaultThemeRenderContext";
 import { JSX } from "../../../../utils";
 import { DeclarationReflection, ReferenceReflection } from "../../../../models";
@@ -17,7 +17,7 @@ export function member(context: DefaultThemeRenderContext, props: DeclarationRef
             <a id={props.anchor} class="tsd-anchor"></a>
             {!!props.name && (
                 <h3 class="tsd-anchor-link">
-                    {renderFlags(props.flags, props.comment)}
+                    {context.reflectionFlags(props)}
                     <span class={classNames({ deprecated: props.isDeprecated() })}>{wbr(props.name)}</span>
                     {anchorIcon(context, props.anchor)}
                 </h3>